added user restrictions and updated tests

+++ project/home/views.py
@@ -2,16 +2,15 @@
 #### imports ####
 #################
 
-from flask import flash, redirect, render_template, request, session, url_for, Blueprint, jsonify, make_response
+from flask import flash, redirect, render_template, request, url_for, Blueprint, jsonify, make_response
 from project import db
 from project.models import Restaurant, User, MenuItem
 
-from flask import session as login_session
+from flask import session as login_session # store values in it for the longgevity of the user in session used for showlogin
 import random
-import string
+import string # used to create sudo random string to identify each session used for showlogin
 from oauth2client.client import flow_from_clientsecrets  # stores client id , secrets
 from oauth2client.client import FlowExchangeError
-from oauth2client.client import AccessTokenCredentials
 import httplib2
 import json
 import requests
@@ -75,7 +74,7 @@ def gconnect():
 
     try:
         # Upgrade the authorization code into a credentials object
-        oauth_flow = flow_from_clientsecrets('node.js', scope='')
+        oauth_flow = flow_from_clientsecrets('client_secret.json', scope='')
         oauth_flow.redirect_uri = 'postmessage'
         credentials = oauth_flow.step2_exchange(code)
     except FlowExchangeError:
@@ -321,17 +320,20 @@ def restaurantsJSON():
 def showRestaurants():
     # restaurants = Restaurant.query.all()
     restaurants = db.session.query(Restaurant).order_by(Restaurant.name.asc())
-    return render_template('restaurants.html', restaurants=restaurants)
+    if 'username' not in login_session:  # render diff templates based on if user is logged in or not
+        return render_template('publicrestaurants.html', restaurants=restaurants)
+    else:
+        return render_template('restaurants.html', restaurants=restaurants)
 
 
 # Create a new restaurant
-
-
 @home_blueprint.route('/restaurant/new/', methods=['GET', 'POST'])
 def newRestaurant():
+    if 'username' not in login_session:  # verify if user is logged in
+        return redirect('/login')
     if request.method == 'POST':
         newRestaurant = Restaurant(
-            name=request.form['name'], user_id=1)
+            name=request.form['name'], user_id=login_session['user_id'])  # add user id to know who created the restaurant
         newRestaurant.save()
         flash('New Restaurant %s Successfully Created' % newRestaurant.name)
         return redirect(url_for('home.showRestaurants'))
@@ -343,9 +345,11 @@ def newRestaurant():
 
 @home_blueprint.route('/restaurant/<int:restaurant_id>/edit/', methods=['GET', 'POST'])
 def editRestaurant(restaurant_id):
-
-    editedRestaurant = db.session.query(
-        Restaurant).filter_by(id=restaurant_id).one()
+    if 'username' not in login_session:
+        return redirect('/login')
+    editedRestaurant = db.session.query(Restaurant).filter_by(id=restaurant_id).one()
+    if editedRestaurant.user_id != login_session['user_id']:
+        return "<script> function myFunction() {alert('You are not authorized to edit this restaurant. Please create your own restaurant in order to delete.');}</script><body onload='myFunction()''>"
     if request.method == 'POST':
         if request.form['name']:
             editedRestaurant.name = request.form['name']
@@ -361,11 +365,13 @@ def editRestaurant(restaurant_id):
 
 @home_blueprint.route('/restaurant/<int:restaurant_id>/delete/', methods=['GET', 'POST'])
 def deleteRestaurant(restaurant_id):
-
+    if 'username' not in login_session:
+        return redirect('/login')
     restaurantToDelete = db.session.query(
         Restaurant).filter_by(id=restaurant_id).one()
     # alert messages tro let user know hes not authorized since he didnt create it, used in create, edit, delete functions
-
+    if restaurantToDelete.user_id != login_session['user_id']:
+        return "<script> function myFunction() {alert('You are not authorized to delete this restaurant. Please create your own restaurant in order to delete.');}</script><body onload='myFunction()''>"
     if request.method == 'POST':
         restaurantToDelete.delete()
         flash('%s Successfully Deleted' % restaurantToDelete.name)
@@ -380,17 +386,24 @@ def deleteRestaurant(restaurant_id):
 @home_blueprint.route('/restaurant/<int:restaurant_id>/menu/')
 def showMenu(restaurant_id):
     restaurant = db.session.query(Restaurant).filter_by(id=restaurant_id).one()
-
+    creator = getUserInfo(restaurant.user_id)  # who created the menu
     items = db.session.query(MenuItem).filter_by(
         restaurant_id=restaurant_id).all()
     # check who created the file and render template based on that
-    return render_template('menu.html', items=items, restaurant=restaurant)
+    if 'username' not in login_session or creator.id != login_session['user_id']:
+        return render_template('publicmenu.html', items=items, restaurant=restaurant, creator=creator)
+    else:
+        return render_template('menu.html', items=items, restaurant=restaurant, creator=creator)
 
 
 # Create a new menu item
 @home_blueprint.route('/restaurant/<int:restaurant_id>/menu/new/', methods=['GET', 'POST'])
 def newMenuItem(restaurant_id):
+    if 'username' not in login_session:
+        return redirect('/login')
     restaurant = db.session.query(Restaurant).filter_by(id=restaurant_id).one()
+    if login_session['user_id'] != restaurant.user_id:
+        return "<script>function myFunction() {alert('You are not authorized to add menu items to this restaurant. Please create your own restaurant in order to add items.');}</script><body onload='myFunction()''>"
     if request.method == 'POST':
         newItem = MenuItem(name=request.form['name'], description=request.form['description'], price=request.form[
                            'price'], course=request.form['course'], restaurant_id=restaurant_id, user_id=restaurant.user_id)
@@ -403,10 +416,15 @@ def newMenuItem(restaurant_id):
 # Edit a menu item
 
 
-@home_blueprint.route('/restaurant/<int:restaurant_id>/menu/<int:menu_id>/edit', methods=['GET', 'POST'])
+@home_blueprint.route('/restaurant/<int:restaurant_id>/menu/<int:menu_id>/edit/', methods=['GET', 'POST'])
 def editMenuItem(restaurant_id, menu_id):
+    if 'username' not in login_session:
+        return redirect('/login')
     editedItem = db.session.query(MenuItem).filter_by(id=menu_id).one()
     restaurant = db.session.query(Restaurant).filter_by(id=restaurant_id).one()
+    if login_session['user_id'] != restaurant.user_id:
+        return "<script>function myFunction() {alert('You are not authorized to edit menu items to this restaurant. Please create your own restaurant in order to edit items.');}</script><body onload='myFunction()''>"
+
     if request.method == 'POST':
         if request.form['name']:
             editedItem.name = request.form['name']
@@ -424,10 +442,15 @@ def editMenuItem(restaurant_id, menu_id):
 
 
 # Delete a menu item
-@home_blueprint.route('/restaurant/<int:restaurant_id>/menu/<int:menu_id>/delete', methods=['GET', 'POST'])
+@home_blueprint.route('/restaurant/<int:restaurant_id>/menu/<int:menu_id>/delete/', methods=['GET', 'POST'])
 def deleteMenuItem(restaurant_id, menu_id):
+    if 'username' not in login_session:
+        return redirect('/login')
     restaurant = db.session.query(Restaurant).filter_by(id=restaurant_id).one()
     itemToDelete = db.session.query(MenuItem).filter_by(id=menu_id).one()
+    if login_session['user_id'] != restaurant.user_id:
+        return "<script>function myFunction() {alert('You are not authorized to delete menu items to this restaurant. Please create your own restaurant in order to delete items.');}</script><body onload='myFunction()''>"
+
     if request.method == 'POST':
         itemToDelete.delete()
         flash('Menu Item Successfully Deleted')
@@ -449,3 +472,27 @@ def upload(restaurant_id, menu_id):
         # SHOULD USE IN YOUR EDITMENUITEM TEMPLATE
         return render_template(
             'upload.html',  restaurant=restaurant, item=item)  # item represents the item we want to edit
+
+@home_blueprint.errorhandler(404)
+def not_found(error):
+    return render_template('404.html')
+
+# error example code
+
+
+@home_blueprint.route('/slashboard/')
+def slashboard():
+    restaurants = db.session.query(Restaurant).order_by(Restaurant.name.asc())
+    try:
+        return render_template('publicrestaurants.html', restaurants=restaurants)
+    except Exception as error:
+        return render_template('404.html', error=error)
+
+
+@home_blueprint.route('/request-info/')
+def request_info():
+    # Get location info using https://freegeoip.net/
+    geoip_url = 'http://freegeoip.net/json/{}'.format(request.remote_addr)
+    client_location = requests.get(geoip_url).json()
+    return render_template('info.html',
+                           client_location=client_location)
